requirements_38.txt
streamlit==1.10.0
streamlit_ace==0.1.1
streamlit_option_menu==0.3.2
streamlit_pandas_profiling==0.1.3
boto3==1.20.29
imbalanced_learn==0.8.1
imblearn==0.0
lightgbm==3.3.1
matplotlib==3.3.2
mlflow==1.25.0
numpy==1.19.5
pandas==1.4.1
pandas_profiling==3.1.0
phik==0.12.0
Pillow==9.2.0
plotly==4.14.3
scikit_learn==1.1.2
scikit_plot==0.3.7
scipy==1.5.2
seaborn==0.11.2
shap==0.39.0
statsmodels==0.12.0
sweetviz==2.1.3
Unidecode==1.3.4
xgboost==1.5.1
yellowbrick==1.3.post1
